refac(rust): put files into subfolders, move preferences logic into own file
